Restructure branch to new folder layout
Moved components into new folder structure, and deleted old, unused
componentsfiles.

Per Inge Vaaje <60141320+PerIngeVaaje@users.noreply.github.com>
Øyvind Edvardsen
Kent Mossbaeck
Mikael Nordberg
