+++ libs/pxweb2-ui/src/index.ts
@@ -1,2 +1,1 @@
-export * from './lib/Button';
-export * from './lib/pxweb2-ui';
+export * from './lib/components/Button/Button';
