--- libs/pxweb2-ui/src/lib/Icons/Pencil.module.scss
@@ -1,5 +1,0 @@
-.pencil {
-  width: 1.5rem;
-  height: 1.5rem;
-  flex-shrink: 0;
-}
