--- libs/pxweb2-ui/src/lib/Icons/Pencil.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import Pencil from './Pencil';
-
-describe('Pencil', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<Pencil />);
-    expect(baseElement).toBeTruthy();
-  });
-});
