+++ libs/pxweb2-ui/src/lib/components/Icon/Icon.module.scss
