+++ libs/pxweb2-ui/src/lib/components/Icon/Icon.spec.tsx
