--- libs/pxweb2-ui/src/lib/pxweb2-ui.module.scss
@@ -1,7 +1,0 @@
-/*
- * Replace this with your own classes
- *
- * e.g.
- * .container {
- * }
-*/
