--- libs/pxweb2-ui/src/lib/pxweb2-ui.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import Pxweb2Ui from './pxweb2-ui';
-
-describe('Pxweb2Ui', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<Pxweb2Ui />);
-    expect(baseElement).toBeTruthy();
-  });
-});
